feat(ColorConverter, PointConverter): Color와 Point를 변환하는 util성 함수들이 모인 파일 생성
