+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -10,14 +10,14 @@ import androidx.core.view.children
 import domain.OmokGame
 import domain.OmokGameState
 import domain.board.Board
-import domain.stone.Color
-import domain.stone.Point
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var boards: List<ImageView>
-    private lateinit var omokDBAdapter: OmokDBAdapter
     private lateinit var omokGame: OmokGame
+    private val omokDBAdapter: OmokDBAdapter by lazy {
+        OmokDBAdapter(OmokDBHelper(this))
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,11 +28,9 @@ class MainActivity : AppCompatActivity() {
             .flatMap { it.children }
             .filterIsInstance<ImageView>()
             .toList()
-        omokDBAdapter = OmokDBAdapter(OmokDBHelper(this))
         omokGame = OmokGame(omokDBAdapter.getStones())
 
         printBoard(omokGame.board)
-
         boards.forEachIndexed { index, view ->
             view.setOnClickListener {
                 placeStone(index)
@@ -44,7 +42,7 @@ class MainActivity : AppCompatActivity() {
     private fun printBoard(currentBoard: Board) {
         omokDBAdapter.getStones()
         currentBoard.placedStones.forEach { placedStone ->
-            boards[placedStone.point.toIndex()].setImageResource(placedStone.color.toResourceId())
+            boards[placedStone.point.toIndex()].setImageResource(placedStone.color.toStoneImage())
         }
     }
 
@@ -54,6 +52,18 @@ class MainActivity : AppCompatActivity() {
             omokGame.placeStone(::printBoard) { _, _ -> clickedIndex.toPoint() }
             printBoard(omokGame.board)
         }
+    }
+
+    private fun clearBoard() {
+        boards.forEach { board ->
+            board.setImageBitmap(null)
+        }
+    }
+
+    private fun initOmokGames() {
+        omokDBAdapter.deleteStones()
+        omokGame = OmokGame()
+        clearBoard()
     }
 
     private fun checkEnded() {
@@ -68,26 +78,9 @@ class MainActivity : AppCompatActivity() {
                 ),
                 Toast.LENGTH_SHORT
             ).show()
-            omokDBAdapter.deleteStones()
-            omokGame = OmokGame()
-            clearBoard()
+            initOmokGames()
         }
     }
-
-    private fun clearBoard() {
-        boards.forEach { board ->
-            board.setImageBitmap(null)
-        }
-    }
-
-    private fun Color.toResourceId(): Int {
-        return when (this) {
-            is Color.Black -> R.drawable.custom_black_stone
-            is Color.White -> R.drawable.custom_white_stone
-        }
-    }
-
-    private fun Point.toIndex(): Int = (RANGE_UNIT * (y - 1)) + (x - 1)
 
     companion object {
         private const val WINNING_MESSAGE = "우승자는 %s입니다."
